requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0
# Deep Learning
torch>=1.9.0
torchvision>=0.10.0
# Embeddings
sentence-transformers>=2.2.0
transformers>=4.10.0
# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0
networkx>=2.6.0
# Optimization
pymoo>=0.5.0
optuna>=2.10.0
# Utilities
pyyaml>=6.0
tqdm>=4.62.0
joblib>=1.1.0
colorama>=0.4.4
# Logging
tensorboard>=2.7.0
wandb>=0.12.0
# Testing
pytest>=6.2.0
pytest-cov>=3.0.0
# Optional: For advanced embeddings
faiss-cpu>=1.7.0
annoy>=1.17.0
